v0.4 changes for slower crystal startup. added initial qc test code
