+++ test/host/clocks.pio
@@ -1,0 +1,35 @@
+/*
+ * Project: pico9918
+ *
+ * Copyright (c) 2024 Troy Schrapel
+ *
+ * This code is licensed under the MIT license
+ *
+ * https://github.com/visrealm/pico9918
+ *
+ */
+
+.program clock
+    pull block
+.wrap_target
+    set pins, 1
+    mov x, osr
+onDelay:
+    jmp x-- onDelay
+    set pins, 0
+    mov x, osr
+offDelay:
+    jmp x-- offDelay
+.wrap
+
+
+% c-sdk {
+
+void clock_program_init(PIO pio, uint sm, uint offset, uint pin) {
+  pio_gpio_init(pio, pin);
+  pio_sm_set_consecutive_pindirs(pio, sm, pin, 1, true);
+  pio_sm_config c = clock_program_get_default_config(offset);
+  sm_config_set_set_pins(&c, pin, 1);
+  pio_sm_init(pio, sm, offset, &c);
+}
+%}
